+++ frb_codegen/src/library/codegen/generator/codec/sse/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod lang;
 pub(crate) mod misc;
 pub(crate) mod ty;
+mod lockable;
